Clean up the solution items

+++ Azofe.Core/Id.cs
@@ -2,19 +2,19 @@ namespace Azofe.Core;
 
 public sealed class Id: ValueObject {
 
-	public Id(long value) {
+	public Id(int value) {
 		if(!IsValid(value))
 			throw new ArgumentException("The identifier must be a number greater than zero.", nameof(value));
 		Value = value;
 	}
 
-	public static implicit operator Id(long value) => new(value);
+	public static implicit operator Id(int value) => new(value);
 
-	public static implicit operator long(Id id) => id.Value;
+	public static implicit operator int(Id id) => id.Value;
 
-	public long Value { get; }
+	public int Value { get; }
 
-	public static bool IsValid(long value) => value > 0;
+	public static bool IsValid(int value) => value > 0;
 
 	protected override IEnumerable<object?> GetEqualityComponents() {
 		yield return Value;
